Combine `AssignTaskCommand` and `UnassignTaskCommand`...
to reduce code duplication
